requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.10.0
django-filter==23.2
django-cors-headers==3.14.0
openpyxl==3.1.2
Pillow==10.0.0
pytest==7.4.0
pytest-django==4.5.2